# backend/contracts/openapi.yaml
{
  "components": {
    "schemas": {
...
      "BatchResults": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/BatchRoundResult"
            },
            "minItems": 1,
            "title": "Items",
            "type": "array"
          }
        },
        "required": [
          "items"
        ],
        "title": "BatchResults",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "bearerFormat": "opaque",
        "description": "Managed high-entropy bearer token. Required scopes are listed in x-required-scopes because this is not an OAuth token endpoint.",
        "scheme": "bearer",
        "type": "http",
        "x-available-scopes": [
          "practice:generate",
          "practice:read",
          "practice:write",
          "reviews:read",
          "reviews:write",
          "words:export",
          "words:read",
          "words:write"
        ]
      },
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      },
      "TrustedProxyUser": {
        "description": "Only accepted from TRUSTED_PROXY_CIDRS after the proxy strips client-supplied headers.",
        "in": "header",
        "name": "X-Forwarded-User",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Single-user word memory API. Browser requests use a trusted reverse-proxy identity; external Skills use scoped bearer tokens.",
    "title": "Word Memory Assistant API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/api-clients": {
      "get": {
        "operationId": "list_clients_api_v1_api_clients_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Clients",
        "tags": [
          "api-clients"
        ]
      },
      "post": {
        "operationId": "create_client_api_v1_api_clients_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApiClientCreateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Create Client",
        "tags": [
          "api-clients"
        ]
      }
    },
    "/api/v1/api-clients/{client_id}": {
      "delete": {
        "operationId": "disable_client_api_v1_api_clients__client_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "client_id",
            "required": true,
            "schema": {
              "title": "Client Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Disable Client",
        "tags": [
          "api-clients"
        ]
      },
      "patch": {
        "operationId": "update_client_api_v1_api_clients__client_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "client_id",
            "required": true,
            "schema": {
              "title": "Client Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApiClientUpdateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Client",
        "tags": [
          "api-clients"
        ]
      }
    },
...
    "/api/v1/daily-table/generate": {
      "post": {
        "operationId": "generate_api_v1_daily_table_generate_post",
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/StrategyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "AUTH_REQUIRED"
          },
          "403": {
            "description": "FORBIDDEN_SCOPE"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "TrustedProxyUser": []
          }
        ],
        "summary": "Generate",
        "tags": [
          "practice"
        ],
        "x-required-scopes": [
          "practice:generate"
        ]
      }
    },
    "/api/v1/practice-review-rounds/{round_id}": {
      "get": {
        "operationId": "get_round_api_v1_practice_review_rounds__round_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "round_id",
            "required": true,
            "schema": {
              "title": "Round Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "AUTH_REQUIRED"
          },
          "403": {
            "description": "FORBIDDEN_SCOPE"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "TrustedProxyUser": []
          }
        ],
        "summary": "Get Round",
        "tags": [
          "practice"
        ],
        "x-required-scopes": [
          "practice:read"
        ]
      }
    },
    "/api/v1/practice-review-rounds/{round_id}/items/{item_id}/result": {
      "put": {
        "operationId": "put_result_api_v1_practice_review_rounds__round_id__items__item_id__result_put",
        "parameters": [
          {
            "in": "path",
            "name": "round_id",
            "required": true,
            "schema": {
              "title": "Round Id",
              "type": "integer"
            }
          },
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoundResult"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "AUTH_REQUIRED"
          },
          "403": {
            "description": "FORBIDDEN_SCOPE"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "TrustedProxyUser": []
          }
        ],
        "summary": "Put Result",
        "tags": [
          "practice"
        ],
        "x-required-scopes": [
          "reviews:write"
        ]
      }
    },
    "/api/v1/practice-review-rounds/{round_id}/results": {
      "put": {
        "operationId": "put_results_api_v1_practice_review_rounds__round_id__results_put",
        "parameters": [
          {
            "in": "path",
            "name": "round_id",
            "required": true,
            "schema": {
              "title": "Round Id",
              "type": "integer"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchResults"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "AUTH_REQUIRED"
          },
          "403": {
            "description": "FORBIDDEN_SCOPE"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "TrustedProxyUser": []
          }
        ],
        "summary": "Put Results",
        "tags": [
          "practice"
        ],
        "x-required-scopes": [
          "reviews:write"
        ]
      }
    },
    "/api/v1/practice-sessions": {
      "get": {
        "operationId": "list_sessions_api_v1_practice_sessions_get",
        "parameters": [
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Size",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "created_by_actor_type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Created By Actor Type"
            }
          },
          {
            "in": "query",
            "name": "generated_from",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Generated From"
            }
          },
          {
            "in": "query",
            "name": "generated_to",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Generated To"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "AUTH_REQUIRED"
          },
          "403": {
            "description": "FORBIDDEN_SCOPE"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "TrustedProxyUser": []
          }
        ],
        "summary": "List Sessions",
        "tags": [
          "practice"
        ],
        "x-required-scopes": [
          "practice:read"
        ]
      }
    },
    "/api/v1/practice-sessions/{session_id}": {
      "delete": {
        "operationId": "delete_session_route_api_v1_practice_sessions__session_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "expected_version",
            "required": true,
            "schema": {
              "exclusiveMinimum": 0,
              "title": "Expected Version",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "description": "AUTH_REQUIRED"
          },
          "403": {
            "description": "FORBIDDEN_SCOPE"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "TrustedProxyUser": []
          }
        ],
        "summary": "Delete Session Route",
        "tags": [
          "practice"
        ],
        "x-required-scopes": [
          "practice:write"
        ]
      },
      "get": {
        "operationId": "get_session_api_v1_practice_sessions__session_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "AUTH_REQUIRED"
          },
          "403": {
            "description": "FORBIDDEN_SCOPE"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "TrustedProxyUser": []
          }
        ],
        "summary": "Get Session",
        "tags": [
          "practice"
        ],
        "x-required-scopes": [
          "practice:read"
        ]
      },
      "patch": {
        "operationId": "update_session_route_api_v1_practice_sessions__session_id__patch",
        "parameters": [
          {
            "in": "path",
            "name": "session_id",
            "required": true,
            "schema": {
              "title": "Session Id",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SessionUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "AUTH_REQUIRED"
          },
          "403": {
            "description": "FORBIDDEN_SCOPE"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "TrustedProxyUser": []
          }
        ],
        "summary": "Update Session Route",
        "tags": [
          "practice"
        ],
        "x-required-scopes": [
          "practice:write"
        ]
      }
    },
...
    "/api/v1/reviews/today": {
      "get": {
        "operationId": "today_online_reviews_api_v1_reviews_today_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodayReviewResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "description": "AUTH_REQUIRED"
          },
          "403": {
            "description": "FORBIDDEN_SCOPE"
          }
        },
        "security": [
          {
            "BearerAuth": []
          },
          {
            "TrustedProxyUser": []
          }
        ],
        "summary": "Today Online Reviews",
        "tags": [
          "reviews"
        ],
        "x-required-scopes": [
          "reviews:write",
          "practice:read"
        ]
      }
    },
...
  }
}